Move extension traits to separate files
